Rename register/behalf/organisation-name to register/agent/name-of-organisation
