+++ packages/forms-web-app/src/routes/register/agent/name-of-organisation.js
@@ -1,17 +1,17 @@
 const express = require('express');
 
-const organisationNameController = require('../../../controllers/register/behalf/organisation-name');
+const organisationNameController = require('../../../controllers/register/agent/name-of-organisation');
 const { validationErrorHandler } = require('../../../validators/validation-error-handler');
 const {
   rules: organisationNameValidationRules,
-} = require('../../../validators/register/behalf/organisation-name');
+} = require('../../../validators/agent/name-of-organisation');
 
 const router = express.Router();
 
-router.get('/organisation-name', organisationNameController.getOrganisationName);
+router.get('/name-of-organisation', organisationNameController.getOrganisationName);
 
 router.post(
-  '/organisation-name',
+  '/name-of-organisation',
   organisationNameValidationRules(),
   validationErrorHandler,
   organisationNameController.postOrganisationName
